+++ tests/interactive/test-use_targets.R
@@ -1,0 +1,4 @@
+tar_test("use_targets()", {
+  use_targets(path = "targets.Rmd", open = FALSE)
+  expect_true(file.exists("targets.Rmd"))
+})